+++ tests/__fixtures__/plugin-exported-definition/index.js
@@ -1,12 +1,11 @@
 const path = require('path');
-const { string, number, bool } = require('optimal');
 
 const Plugin = require(path.join(__dirname, '../../../packages/core/src/Plugin')).default;
 
 module.exports = class TestPlugin extends Plugin {
-  blueprint() {
+  blueprint({ string, bool, number }) {
     return {
-      foo: string().empty(),
+      foo: string(),
       bar: bool(),
       baz: number(),
     };