+++ tests/__fixtures__/reporter/index.js
@@ -1,10 +1,9 @@
 const path = require('path');
-const { number } = require('optimal');
 
 const Reporter = require(path.join(__dirname, '../../../packages/core/src/Reporter')).default;
 
 module.exports = class TestReporter extends Reporter {
-  blueprint() {
+  blueprint({ number }) {
     return {
       fps: number(),
     };